Use locale-independent mdfind query
Apparently the `kind:Application` type of mdfind query is per-locale (so if you have Japanese added as one of your locales, you could also type `種類:アプリ` and it would work just as well).  This also means that if the user *doesn't* have English set as one of their locales, `kind:Application` doesn't work.